add confirm dialog

+++ src/app/products/list.component.ts
@@ -3,14 +3,17 @@ import { first } from 'rxjs/operators';
 
 import { ProductService } from 'src/app/_services';
 import { Product } from 'src/app/_models';
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
 
 
-@Component({ templateUrl: 'list.component.html' })
+
+@Component({ templateUrl: 'list.component.html' ,
+providers: [ConfirmationService]})
 export class ListComponent implements OnInit {
     Products!: Product[];
+    icon: string | undefined;
 
-    constructor(private Productservice: ProductService,private messageService: MessageService) {}
+    constructor(private Productservice: ProductService,private messageService: MessageService,private confirmationService: ConfirmationService) {}
 
     ngOnInit() {
         this.Productservice.getAll()
@@ -19,15 +22,27 @@ export class ListComponent implements OnInit {
     }
 
     deleteproduct(id: string) {
-        const product = this.Products.find(x => x.id === id);
-        if (!product) return;
-        product.isDeleting = true;
-        this.Productservice.delete(id)
-            .pipe(first())
-            .subscribe(() => this.Products = this.Products.filter(x => x.id !== id));
-            this.messageService.add({severity:'info', summary: 'Record is deleted successully', detail:'record deleted'});
+
+        this.icon = "fa-question-circle";
+        this.confirmationService.confirm({
+            message: 'are you sure to delete record.',
+            header: 'delete confirm',
+           
+            accept: () => {
+                const product = this.Products.find(x => x.id === id);
+                if (!product) return;
+                product.isDeleting = true;
+                this.Productservice.delete(id)
+                    .pipe(first())
+                    .subscribe(() => this.Products = this.Products.filter(x => x.id !== id));
+                    this.messageService.add({severity:'info', summary: 'Record is deleted successully', detail:'record deleted'});
+            },
+            reject: () => {
+            }
+        });
+       
     }
   
-    
+
   
 }
